+++ commands/index.js
@@ -1,6 +1,5 @@
 const {accountingCategories} = require('./constants');
 const callRawApi = require('./call_raw_api');
-const {exchanges} = require('./constants');
 const fetchRequest = require('./fetch_request');
 const {marketPairs} = require('./constants');
 const {peerSortOptions} = require('./constants');
@@ -10,7 +9,6 @@ const {swapTypes} = require('./constants');
 module.exports = {
   accountingCategories,
   callRawApi,
-  exchanges,
   fetchRequest,
   marketPairs,
   peerSortOptions,